Moved rate limiter to services
